[IR] Merged `CFG` trait in `ControlFlow`.
