+++ emma-language/src/main/scala/org/emmalanguage/compiler/lang/cf/ControlFlow.scala
@@ -17,12 +17,18 @@ package org.emmalanguage
 package compiler.lang.cf
 
 import compiler.Common
+import compiler.ir.DSCFAnnotations.continuation
 import compiler.lang.core.Core
+import util.Monoids._
 
-import quiver.Graph
+import cats.instances.all._
+import quiver._
+import shapeless._
 
-/** Backend-related (but backend-agnostic) transformations. */
-private[compiler] trait ControlFlow extends Common with CFG {
+import scala.collection.breakOut
+
+/** Control-flow graph (CFG) construction. */
+private[compiler] trait ControlFlow extends Common {
   self: Core =>
 
   case class FlowGraph[V]
@@ -33,9 +39,82 @@ private[compiler] trait ControlFlow extends Common with CFG {
     data: Graph[V, u.ValDef, Unit]
   )
 
+  /** Control-flow graph analysis (CFG). */
   object ControlFlow {
 
-    /** Delegates to [[CFG.graph]]. */
-    lazy val cfg: u.Tree => FlowGraph[u.TermSymbol] = CFG.graph
+    import API.GraphRepresentation.phi
+    import Core.{Lang => core}
+    import UniverseImplicits._
+
+    private val module = Some(API.GraphRepresentation.ref)
+
+    /**
+     * Control-flow graph analysis (CFA) of a tree.
+     *
+     * == Preconditions ==
+     * - The input `tree` is in DSCF (see [[DSCF.transform]]).
+     *
+     * == Postconditions ==
+     * - The RHS of method parameters is represented by `phi` nodes.
+     * - The graph contains:
+     *   - The number of references for each val / parameter;
+     *   - All nest(ing) relations from outer to inner;
+     *   - All control (flow) dependencies from caller to callee;
+     *   - All data (flow) dependencies from RHS to LHS.
+     */
+    lazy val cfg: u.Tree => FlowGraph[u.TermSymbol] = {
+      api.TopDown.withDefDefs.withBindUses.withBindDefs
+        .inherit { case core.TermDef(encl) => Option(encl) } (last(None))
+        .accumulateWith[Vector[UEdge[u.TermSymbol]]] { // Nesting
+          case Attr.inh(core.BindingDef(lhs, _), Some(encl) :: _) =>
+            Vector(LEdge(encl, lhs, ()))
+          case Attr.inh(core.DefDef(method, _, _, _), Some(encl) :: _) =>
+            Vector(LEdge(encl, method, ()))
+        }.accumulate { // Control flow
+          case core.DefDef(f, _, _, core.Let(_, _, expr)) =>
+            expr.collect { case core.DefCall(None, g, _, _)
+              if api.Sym.findAnn[continuation](g).isDefined =>
+              LEdge(f.asTerm, g.asTerm, ())
+            }.toVector
+        }.accumulateWith[Vector[UEdge[u.TermSymbol]]] { // Data flow
+          case Attr.inh(core.BindingRef(from), Some(to) :: _) if !from.isStatic =>
+            Vector(LEdge(from, to, ()))
+        }.accumulate { // Phi choices
+          case core.DefCall(None, method, _, argss)
+            if api.Sym.findAnn[continuation](method).isDefined => {
+              for ((param, arg) <- method.paramLists.flatten zip argss.flatten)
+                yield param.asTerm -> Vector(arg)
+            }.toMap
+        } (merge).traverseAny.andThen {
+          case Attr(_, choices :: data :: ctrl :: nest :: _, _, vals :: uses :: defs :: _) =>
+            val nestTree = quiver.empty[u.TermSymbol, Unit, Unit]
+              .addNodes(vals.keys.map(x => LNode(x, ()))(breakOut))
+              .addNodes(defs.keys.map(m => LNode(m.asTerm, ()))(breakOut))
+              .safeAddEdges(nest)
+
+            val ctrlFlow = quiver.empty[u.TermSymbol, u.DefDef, Unit]
+              .addNodes(defs.map { case (k, v) => LNode(k.asTerm, v) } (breakOut))
+              .safeAddEdges(ctrl)
+
+            val phiNodes = for {
+              api.ParSym(lhs) <- vals.keys.toStream if choices.contains(lhs)
+              rhs = core.DefCall(module, phi, Seq(lhs.info), Seq(choices(lhs)))
+            } yield LNode(lhs, core.ParDef(lhs, rhs))
+
+            val phiEdges = for {
+              (lhs, args) <- choices.toStream
+              core.Ref(target) <- args if !target.isStatic
+            } yield LEdge(target, lhs, ())
+
+            val dataFlow = quiver.empty[u.TermSymbol, u.ValDef, Unit]
+              .addNodes(vals.map { case (k, v) => LNode(k, v) } (breakOut))
+              .addNodes(phiNodes)
+              .safeAddEdges(data)
+              .safeAddEdges(phiEdges)
+
+            FlowGraph(uses, nestTree, ctrlFlow, dataFlow)
+        }
+    }
   }
+
 }
